\name{internal.oncotree}
\alias{build.parent}
\alias{build.plot}
\alias{formatData}
\title{Undocumented internal functions for the Oncotree package}
\description{These functions should not be called by the user.
}
\usage{
  build.parent(x)
  build.plot(parent)
  formatData(dataset)
}
\keyword{internal}
